[refactor] dto 폴더 명 변경
- dtos -> dto로 변경
